test(eval): integration coverage for the polars pipeline; fixes 2 production bugs
New test_pipeline.py covers _score_outcomes, _aggregate,
_select_misses from measure and _score_trials from tune against
synthetic SearchBatch / WeightedSearchBatch frames.  Twelve
thin behavioural assertions against bespoke tiny datasets
(2-6 rows each); no daemon, no embedding model.

Caught two production bugs:

- measure._select_misses: the gap column was subtracting UInt8
  rank values and only casting to Int16 afterwards.  When
  stripped < full (stripped did BETTER than full), the UInt8
  subtraction wrapped around to a large positive number, so a
  row that should have been filtered out appeared as the
  highest-gap miss.  Fix: cast to Int16 before the subtraction.
- tune._score_trials: the left-join back to the raw trials
  frame used polars' default null-inequality semantics.
  Baseline rows have null alpha/beta/gamma, so every baseline
  rank silently came back null after the join, and the
  reported score_current in tuned-params.json was always 0.0.
  Fix: join with nulls_equal=True.  The real baseline MRR
  surfaces in the smoke test (0.35 vs the previously reported
  0.00); delta drops from an inflated 0.90 to a real 0.65.

+++ packages/rbtr-eval/src/rbtr_eval/measure.py
@@ -219,9 +219,10 @@ def _select_misses(
             values=["rank", "top_file", "top_line", "top_name"],
         )
         .with_columns(
-            (pl.col("rank_stripped").fill_null(11) - pl.col("rank_full").fill_null(11))
-            .cast(pl.Int16)
-            .alias("gap")
+            (
+                pl.col("rank_stripped").fill_null(11).cast(pl.Int16)
+                - pl.col("rank_full").fill_null(11).cast(pl.Int16)
+            ).alias("gap")
         )
         .filter(pl.col("gap") > 0)
         .sort(["gap", "slug", "query_file"], descending=[True, False, False])
